requirements.txt
google-generativeai
langchain
google-search-results
streamlit
python-dotenv
langchain-google-genai 